# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MediQ",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "operationId": "me_api_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/providers/{provider_id}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Get Provider",
        "operationId": "get_provider_api_providers__provider_id__get",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/appointments": {
      "post": {
        "tags": [
          "booking"
        ],
        "summary": "Create Appointment",
        "operationId": "create_appointment_api_appointments_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "booking"
        ],
        "summary": "List Appointments",
        "operationId": "list_appointments_api_appointments_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_VisitOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/appointments/{visit_id}/cancel": {
      "post": {
        "tags": [
          "booking"
        ],
        "summary": "Cancel Appointment",
        "operationId": "cancel_appointment_api_appointments__visit_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "visit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Visit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/visits/{visit_id}/check-in": {
      "post": {
        "tags": [
          "booking"
        ],
        "summary": "Check In Visit",
        "operationId": "check_in_visit_api_visits__visit_id__check_in_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "visit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Visit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/settings": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Settings",
        "operationId": "get_settings_api_admin_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClinicSettingsOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Update Settings",
        "operationId": "update_settings_api_admin_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClinicSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClinicSettingsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/services/{service_id}": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Update Service",
        "operationId": "update_service_api_admin_services__service_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "service_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Service Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/providers": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List All Providers",
        "operationId": "list_all_providers_api_admin_providers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProviderOut"
                  },
                  "type": "array",
                  "title": "Response List All Providers Api Admin Providers Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create Provider",
        "operationId": "create_provider_api_admin_providers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/providers/{provider_id}": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Update Provider",
        "operationId": "update_provider_api_admin_providers__provider_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Provider Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppointmentCreate": {
        "properties": {
          "patient_id": {
            "type": "integer",
            "title": "Patient Id"
          },
          "provider_id": {
            "type": "integer",
            "title": "Provider Id"
          },
          "service_id": {
            "type": "integer",
            "title": "Service Id"
          },
          "slot_id": {
            "type": "integer",
            "title": "Slot Id"
          }
        },
        "type": "object",
        "required": [
          "patient_id",
          "provider_id",
          "service_id",
          "slot_id"
        ],
        "title": "AppointmentCreate"
      },
      "ClinicSettingsOut": {
        "properties": {
          "noshow_grace_minutes": {
            "type": "integer",
            "title": "Noshow Grace Minutes"
          },
          "default_overbook_limit": {
            "type": "integer",
            "title": "Default Overbook Limit"
          },
          "ewma_alpha": {
            "type": "number",
            "title": "Ewma Alpha"
          },
          "walkin_routing_enabled": {
            "type": "boolean",
            "title": "Walkin Routing Enabled"
          }
        },
        "type": "object",
        "required": [
          "noshow_grace_minutes",
          "default_overbook_limit",
          "ewma_alpha",
          "walkin_routing_enabled"
        ],
        "title": "ClinicSettingsOut"
      },
      "ClinicSettingsUpdate": {
        "properties": {
          "noshow_grace_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Noshow Grace Minutes"
          },
          "default_overbook_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Overbook Limit"
          },
          "ewma_alpha": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ewma Alpha"
          },
          "walkin_routing_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Walkin Routing Enabled"
          }
        },
        "type": "object",
        "title": "ClinicSettingsUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
      "LoginResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "provider_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Id"
          },
          "patient_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Id"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "role",
          "user_id"
        ],
        "title": "LoginResponse"
      },
...
      "Page_VisitOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/VisitOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[VisitOut]"
      },
...
      "ProviderCreate": {
        "properties": {
          "department_id": {
            "type": "integer",
            "title": "Department Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "$ref": "#/components/schemas/ProviderKind"
          },
          "room_label": {
            "type": "string",
            "title": "Room Label"
          },
          "shift_start": {
            "type": "string",
            "format": "time",
            "title": "Shift Start"
          },
          "shift_end": {
            "type": "string",
            "format": "time",
            "title": "Shift End"
          },
          "slot_length_min": {
            "type": "integer",
            "title": "Slot Length Min"
          },
          "slot_capacity": {
            "type": "integer",
            "title": "Slot Capacity",
            "default": 1
          },
          "overbook_limit": {
            "type": "integer",
            "title": "Overbook Limit",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "department_id",
          "name",
          "kind",
          "room_label",
          "shift_start",
          "shift_end",
          "slot_length_min"
        ],
        "title": "ProviderCreate"
      },
...
      "ProviderOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "department_id": {
            "type": "integer",
            "title": "Department Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "$ref": "#/components/schemas/ProviderKind"
          },
          "room_label": {
            "type": "string",
            "title": "Room Label"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "shift_start": {
            "type": "string",
            "format": "time",
            "title": "Shift Start"
          },
          "shift_end": {
            "type": "string",
            "format": "time",
            "title": "Shift End"
          },
          "slot_length_min": {
            "type": "integer",
            "title": "Slot Length Min"
          },
          "slot_capacity": {
            "type": "integer",
            "title": "Slot Capacity"
          },
          "overbook_limit": {
            "type": "integer",
            "title": "Overbook Limit"
          }
        },
        "type": "object",
        "required": [
          "id",
          "department_id",
          "name",
          "kind",
          "room_label",
          "is_active",
          "shift_start",
          "shift_end",
          "slot_length_min",
          "slot_capacity",
          "overbook_limit"
        ],
        "title": "ProviderOut"
      },
...
      "ServiceOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "department_id": {
            "type": "integer",
            "title": "Department Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "default_duration_min": {
            "type": "integer",
            "title": "Default Duration Min"
          },
          "prep_time_min": {
            "type": "integer",
            "title": "Prep Time Min"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "department_id",
          "name",
          "default_duration_min",
          "prep_time_min",
          "is_active"
        ],
        "title": "ServiceOut"
      },
      "ServiceUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "default_duration_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Duration Min"
          },
          "prep_time_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prep Time Min"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "ServiceUpdate"
      },
...
      "UserOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "provider_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Id"
          },
          "patient_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "role",
          "is_active"
        ],
        "title": "UserOut"
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "patient",
          "receptionist",
          "provider",
          "admin"
        ],
        "title": "UserRole"
      },
...
      "VisitOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "patient_id": {
            "type": "integer",
            "title": "Patient Id"
          },
          "provider_id": {
            "type": "integer",
            "title": "Provider Id"
          },
          "service_id": {
            "type": "integer",
            "title": "Service Id"
          },
          "slot_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slot Id"
          },
          "source": {
            "$ref": "#/components/schemas/VisitSource"
          },
          "token_no": {
            "type": "string",
            "title": "Token No"
          },
          "status": {
            "$ref": "#/components/schemas/VisitStatus"
          },
          "scheduled_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled Start"
          },
          "checked_in_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked In At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "delay_minutes": {
            "type": "integer",
            "title": "Delay Minutes"
          },
          "delay_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delay Reason"
          },
          "priority_flag": {
            "type": "boolean",
            "title": "Priority Flag"
          },
          "priority_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority Reason"
          },
          "estimated_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Start"
          },
          "estimated_wait_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Wait Min"
          },
          "eta_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eta Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "patient_id",
          "provider_id",
          "service_id",
          "slot_id",
          "source",
          "token_no",
          "status",
          "scheduled_start",
          "checked_in_at",
          "started_at",
          "completed_at",
          "delay_minutes",
          "delay_reason",
          "priority_flag",
          "priority_reason",
          "estimated_start",
          "estimated_wait_min",
          "eta_reason",
          "created_at"
        ],
        "title": "VisitOut",
        "description": "Shared representation of a visit, returned by booking, walk-in, and lifecycle\nendpoints. Never includes patient contact details beyond what the caller's role\nis entitled to \u2014 the public status endpoint uses PublicStatusOut instead."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}